Add files via upload
This is the HTML file for our calculator. Here, we are using the HTML to create the content of the calculator UI. We are including the buttons of a calculator, input fields, display area. In the HTML code, we have used onclickevent, which means that whenever user clicks on any of the buttons then the corresponding operation is performed at the back...